requirement.txt
aiofile==3.8.8
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
amazon-transcribe==0.6.2
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
awscrt==0.16.26
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.34.13
botocore==1.34.13
cachetools==5.3.2
caio==0.9.13
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
datasets==2.15.0
dill==0.3.6
distro==1.9.0
dnspython==2.5.0
dropbox==11.36.2
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.0
flask-sock==0.7.0
frozenlist==1.4.1
fsspec==2023.10.0
google==3.0.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-speech==2.23.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
keyboard==0.13.5
langchain==0.0.354
langchain-community==0.0.13
langchain-core==0.1.14
langchain-openai==0.0.3
langsmith==0.0.83
loguru==0.7.2
lz4==4.3.3
MarkupSafe==2.1.3
marshmallow==3.20.2
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.6.1
packaging==23.2
pandas==2.2.0
pinecone-client==2.2.4
playsound==1.3.0
ply==3.11
proto-plus==1.23.0
protobuf==3.20.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pygame==2.5.2
PyJWT==2.8.0
pyngrok==7.0.5
pypdf==3.17.4
python-dateutil==2.8.2
python-vlc==3.0.20123
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rsa==4.9
s3transfer==0.10.0
scipy==1.10.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.25
stone==3.3.1
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
twilio==8.11.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
webrtcvad==2.0.10
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
